Update F02Handler.cpp
This program has been revised specifically for the FarmBot irrigation system and should only be used in the 100 ml to 850 ml range. First, a water curve (i.e. how many milliliters of water emerge after how many seconds) was recorded. Because this water curve is not linear, a linear approximation was carried out and a measuring range was selected that has the smallest possible error. This measuring range is between 1.25 seconds (~100 ml) and 20 seconds (~880 ml). A linear approximation was carried out for this range using the Quickfit minitool in Origin and the polynomial milliliter = 42.64 ml * t + 18.75 ml was determined. The same was done for the number of pulses of the water flow sensor YF-S201 and the polynomial: pulses = 21.79 pulses * t + 14.64 pulses was determined. To obtain the milliliter quantity per pulse, the two coefficients 42.64 ml and 21.79 pulses were divided by each other. The result is the conversion factor (1.96 ml/pulse). With the conversion factor and the offset of 18.75 ml, the milliliter quantity can now be calculated based on the measured pulses.